+++ SDeMo/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.10.0` - 2026-04-05
+
+- **added** an `explainmodel` function for model explanations
+- **added** a `ModelExplanation` struct for model explanations, with `CeterisParis`, `PartialResponse`, `ShapleyMC`, and `PartialDependence` options
+- **added** a `featureimportance` function for feature importance, supporting `PartialDependence`, `ShapleyMC`, and `PermutationImportance` options
+- **added** a vignette on feature importance
+- **improved** the vignette on model interpretation
+
 ## `v1.9.0` - 2026-03-31
 
 - **added** support for `OccurrencesInterface` [#569]